Prefer restricted items over common items in restricted areas
This came up in the TK instances when I wanted Bottled Nethergon Energy to take precedence over my normal Super Mana Potions - they both have the same mana values, but the default iteration was causing the mana potions to take precedence when building the bests list. This change breaks a tie between the same values by giving precedence to an item that is usually restricted but isn't at the moment, indicating that you're, say, inside an instance where you can use the item.